add log, use int as dao query param

+++ springEcommerceGuli/backend/EcommerceGuli/gulimall-ware/src/main/java/com/yen/gulimall/ware/dao/WareSkuDao.java
@@ -17,5 +17,5 @@ public interface WareSkuDao extends BaseMapper<WareSkuEntity> {
     void addStock(@Param("skuId") Long skuId, @Param("wareId") Long wareId, @Param("skuNum") Integer skuNum);
 
     // https://youtu.be/JRPBm5sK4Gg?t=469
-    long getSkuStock(@Param("skuId") Long skuId);
+    long getSkuStock(@Param("skuId") int skuId);
 }
